Sync provisioning and kamailio lcr rules sync
It will be executed with no changes on new 
deploys and new upgrades since the changes are 
already edited in previous revs. 

For those ones which already applied the original
revs this one will fix the asyncs but still won't
delete the inconsistent entries delted in 
provisioning and still present in kamailio. 
That's for another rev.