MT#3963 Execute all db-schema scripts in sequential order sorted by ID revision number
Without this change the following ordering is present:

 3.0:   http://paste.mgm.sipwise.com/show/513/
  vs
 trunk: http://paste.mgm.sipwise.com/show/514/

So /usr/share/ngcp-db-schema/db_scripts/base/0070_not_replicated.up
is executed between 0060_* and 0080_* on ngcp 3.0 but after 6746.up
(the last script without the _not_replicated.up suffix) and
together with all the other *_not_replicated scripts before the
next bunch of scripts (being /usr/share/ngcp-db-schema/db_scripts/diff/).

This adds a short penalty on execution time because we re-run
ngcp-check-rev-applied for the *_not_replicated scripts but it's
the only way to get it right if we want to execute all db-schema
scripts in sequential order sorted by their ID revision number.