MT#4973 Check node specific execution of *_not_replicated.up scripts
Scripts that need to be executed on all hosts (the ones with
filename *_not_replicated.up) but have been executed on sp1
already are listed as "already executed" on sp2 if the
ngcp-check-rev-applied call doesn't use the node specific check.

So generate a list of *_not_replicated.up files and run
node specific checks while generating the list of missing
revision scripts.