Update patch set 2
Patch Set 2:

If the *-pro package is valid then its Provides satisfies the dependency. If it's not satisfied yet though then a specific package should be chosen to satisfy the dependency, a virtual package isn't good for that because it's unclear which package should be chosen. The *-ce packages are the "most-available" ones, so they should be a safe choice. Does this clarify it for you?

Patch-set: 2
Reviewer: Michael Prokop <1000000@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=0