Update patch set 1
Patch Set 1:

as discussed with irina, modifying values in provisioning.voip_preferences.data_type could be dangerous because this changes might propagate into tables of the kamailio schema, so we abandon this for now. Processing attribute values (update_item) in Role\API\Preferences.pm is rigorously based on using key fields instead of id values. In order to not break backward compatibility, these measures have to be discussed in detail.

Patch-set: 1