Update patch set 1
Patch Set 1:

the unique&replace query is really nice, but requires to maintain that additional table. i assume the uploaded file is supposed to replace existing data entirely, so the concerns are about duplicate rows in that uploaded file. Since you are iterating the csv in Utils/Billing.pm anyway, wouldn't it be easier to include the row unique check in the loop using a perl hash?

Patch-set: 1