Update patch set 1
Patch Set 1:

We don't know which rows will have duplicate in the DB. But your comment made me to think about situation when we have received file with the record, which has is present already in the unique and raw tables, and in the raw table has lower id. it is possible, that we will need to replace the unique row in such case.

Patch-set: 1