Update patch set 1
Patch Set 1:

Reply on:

Gerhard Jungwirth
May 18 5:20 PM

Patch Set 1:

some thoughts: 1. about the minimal id, I don't know, but if it does, then lgtm. 2. about the key: are we using type at all, what about the billing zone?

we don't have type in the csv (but really it is configurable) but we have it in the table, with "call" value as default, other possible is "sms"

About billing_zone_id:
billing_fees it is a main table, billing_zones, as a data normalization, just keeps name and detailed description for the zone prefix, defined in the billing_fees.

Patch-set: 1