Update patch set 3
Patch Set 3:

I suppose we should also allow value NULL "use domain default" on the profile level as well. At least this is the default state until it is changed by the user and we need to give user an ability to change it back to that state. Right?

Patch-set: 3