Update patch set 3
Patch Set 3:

the number of cash balances and other fields here is fixed.

transforming into an eav data model is planned, but not part of this change.

this is a refactoring topic, which affects the majority of existing cdr columns (beginning with gp0-99?) anyway not only those new ones here.

Patch-set: 3
Label: Code-Review=0