Update patch set 3
Patch Set 3:

we can leave the current cdr fields be and just implement the new addition properly, and cdr cleanup as the next refactoring step. no need to refactor current accounting.cdr state right now.
but adding that block of purely repetitive columns to accounting.cdr now just to remove it after is not good imho.
Although I understand that there are other factors, like how everything is in "hurry" to deliver that specific feature and how much time rate-o-mat + mediator adjustment will take.

the main issue with adding such change now and refactoring it later is someone may start using the newly added column and when such refactoring will take place everything will be just more complex and messy (and overall more human time spent on this).

to be clear: I am not talking about refactoring now, just to use proper relations and schema for this specific change now, so all the current logic works with the same accounting.cdr and the new logic that relates on this change uses new tables.

my review for this change is:

db concept: -2
syntax (except the trailing spaces): +2

Patch-set: 3
Label: Code-Review=0