Update patch set 4
Patch Set 4:

there will be a accountin.cdr_details and accounting.cdr_details_attributes table once in a time:

>accountin.cdr_details:
>(id,cdr_id,partitioning_timestamp,attribute_id,value)

>accountin.cdr_detail_attributes:
>(id,name,max_repetitions)

anyway, we are in doubt if this is a good idea at all. cdr is in the magnitude of 100mio. rows, so cdr_details will be 80*100mio. records. changeing the structure will affect a multitude of components, while its still under question if simple selects on the eav will keep beeing fast when partitioning is enabled once.

i would like to avoid to further blow up the related rate-o-mat change and test cases here to cover even more tickets. furthermore we will have to delay with melita even longer.

btw. there is no difference if the refactoring one day has to affect 40 existing cdr columns or 80.

Patch-set: 4