Update patch set 2
Patch Set 2:

Kirill, i know there may be inconsistencies on the existing platforms but it is not possible to fix them automatically for them. But it's quite seldom that somebody has many records in voip_allowed_ip_groups.

The problem with this .up script is that it is shifting all voip_usr_preferences values by 100+max(id) from voip_allowed_ip_groups and then also changing id for the records from voip_allowed_ip_groups where id <= 100 to greater numbers so after update we don't by which offset all values in voip_usr_preferences were moved. There is detailed explanation in the linked ticket.
In short, the offset can be found from the binlog or with help of the backup. Then we need to update the preferences values back to groupd_id-offset where groupd_id-offset > 100. So no, it can't be done without manual intervention.

Patch-set: 2