Update patch set 2
Patch Set 2:

is nature of this column is purely temporal (e.g. session based) or it stores some data permanently ?

if it is purely temporal - ok from me to drop it.
if it stores some persistent data - not ok. we keep it and mark as "deprecated since" (the logic will be introduced soon) to be cleaned up along with other obsolete columns on a timed and predicted basis. reason for that is we cannot restore (revert) such "drop column" operations and it may be very cruel in certain circumstances with permanent data storage columns.

Patch-set: 2
Label: Code-Review=0