Update patch set 2
Patch Set 2: Code-Review+1

why not "drop index if exists", is it to make the execution safe for nonexistent accounting db, because drop index accouningt.blah_idx is not possible?

ok and now you switche to not replicated, but drop index shouldnt be slow? or i guess it is slow..

Patch-set: 2