Update patch set 2
Patch Set 2:

There is not possibility to write "drop indx if exist".

There is possibility to write "if select from information schema ... drop index .. " while kirill tried to create and universal procedure for further index dropping in the future. while missed the fact that it must be not replicated.


dropping is usually fast, as you mark index as dropped and do not erase data physically.

Patch-set: 2