Update patch set 4
Patch Set 4:

the intention of this .up details is to have it the exact same way as we did it for cdr. it already prooved: just today UPC came up with another word "we want xtension numbers". with this approach we can add it now easily, without any alter table, so no blocking stmt's which implies safe to backport.

Patch-set: 4