Update patch set 4
Patch Set 4: Code-Review-1

Please don't add new columns in cdr in the middle of the table. They have to be added in the end. Otherwise it is impacting upgrades on PRO. AFAIR we had the following issue in the past:
1) NGCP is updated on inactive node. At this time active node is working with old DB schema, because script is not replicated. So far so good.
2) inactive node is promoted to active and starts serving new calls...
3) the replicate on new inactive node running old version breaks
If the new columns however are added at the end, the replication can continue, because we allow non-LOSSY conversions... Please correct me if I am wrong.
P.S. We also had a different kind of problems in the past when the CDR schema change was done in the replicated script.. This was not better.

Patch-set: 4
Label: Code-Review=-1