Update patch set 2
Patch Set 2:

ah, I see now where the BIGINT comes from :)

"If set to 1, Kamailio expects to deal with BIGINT type on database columns for expires and last modified values. It allows to handle better the daylight time adjustments. If set to 0, those columns are expected to be on default type 'DATETIME'. When it is 1, the database columns types have to be changed manually to 'BIGINT'.

Default value is “0”.

Patch-set: 2