TT#24408 Deleting DB user 'export' again for PRO/Carrier systems installed in mr5.[0-3]
The previous fix was backported to mr4.5 and mr3.8 LTS while mr5.[0-3] was untouched.
It resulted to the fact that systems upgraded via release where grants.yml has user 'export'
will have it restored. Also we need to care about Carrier's 3308 here so we have to use _not_replicated.

The current fix will be included into mr5.5.1+ only,
so it is safe to use "IF EXISTS" as in upgrade
mr4.5->mr5.5 it will be executed in new MariaDB 10.1

Change-Id: I5f1612393aea2676f96d6824a9ab641d29bb8407