Update patch set 1
Patch Set 1:

it's meant to contain variables exposed from config.yml that needs to be available in the DB (e.g. used in selects,views). and exported into the db via "ngcpcfg commit"
it is currently required for the general.timezone variable to be exported into the DB and used in the current v_subscriber_timezone, v_contract_timezone, v_reseller_timezone views as a fallback to the currently defined ngcp timezone.

another option would be to have a custom /etc/mysql/vars.cnf with user defined variables.
so on ngcpcfg apply in that case the vars.cnf template is updated and mysql set ... varname is triggered to update the current values.

Patch-set: 1
Label: Code-Review=0