Update patch set 5
Patch Set 5:

Well I tested it throughly and it definitely works as expected, not sure what kind of additional tests are required but I want to have it merged as soon as possible, otherwise we will have more customers upgraded to new release and with non-chunked huge data blobs in firmwares.
P.S. Provided with full details of how it was tested in both this review and the related task.

Patch-set: 5