Update patch set 1
Patch Set 1:

ngcp.db_schema does not record wether it was a replicated or not_replicated. to my udnerstanding, upgrade does 2 stages:

1. run bunch of replciated .up's
2. then run bunch of not replicated .up's

this change moves one not_replicated to the replicated ones, so it's run earlier and prevents the issue when interleaving steps 1. and 2. on nodes (or not interleaving them).

in the end the only difference between upgraded systems prior to this is that the order of the .up' recorded in ngcp.db_schema is different.

Patch-set: 1
Label: Code-Review=0