TT#32220 Cleanup related with permanent use of "policy-rc.d" and migration to systemd
For a long time our installation and upgrade uses a temporary "policy-rc.d" to
prevent to [re]start services in the middle of the installation, so some uses of
"invoke-rc.d" (which defers to "policy-rc.d") can cause changes in behaviour.

Upon review because of this change, we found out that the restart of kamailio
when updating ngcp-db-schema didn't happen for a long time because of the use of
"policy-rc.d" that we were using during installations of upgrades in the last
few months/years, but besides this, the code is obsolete since quite a while
ago.

So we drop it completely now.

Change-Id: I76eae2fbda340a1021f10d28f95e76b35ea92c90