TT#36660 new billing mappings tables, procedures, views
to optimize the heavyweight fraud events query (and similar
places), the billing.billing_mappings is refactored into 2
new tables that will hold a scan-line rendered representation
of the former billing_mapping records known with their
(nullable) start and (nullable) stop cols.

acessing this new tables (eg. insert a billing mapping to
switch a contract's billing_profile, or query for the
contracts current billing profile) is not supposed to be
done explicitly, but by the stored sql proc provided:

- "billing.set_profile_network":
set/add a single billing profile.

- "billing.set_profile_networks_from_package":
set/add a bunch of profiles from a profile package
at once.

- "billing.get_billing_profile":
get a contract's billing profile id at the given epoch.

- "billing.get_billing_profile_ipv4":
get a contract's billing profile id for the given IPv4
string at the given epoch.

- "billing.get_billing_profile_ipv6":
get a contract's billing profile id for the given IPv6
string at the given epoch.

The enhanced fraud queries make use of a view that
can be used to get the current billing profile id:

- "billing.v_actual_billing_profiles":
View showing all contracts' billing_profile_id at the
current time. In contrast to the procs, the view can
be used nicely in joins of stmts.

Finally the .up scripts also cover copying+transforming
the existing records in billing.billing_mappings. the old
table and records are left for now but are subject to be
dropped soon.

Change-Id: Ib5ddd620b2b08ed7db4b6448a44931d37ea5b06a