Update patch set 1
Patch Set 1:

it is already in mr5.5.5 so all customers upgraded to mr5.5.5 receive it.
as for mr5.5.4 even if we hotfix it existing customers will not automatically receive it after the db-schema package update, but at least customers like Netcologne or others where there are noticeable performance issues because of cdrs can be upgraded with the up script by our ops.

Patch-set: 1
Label: Code-Review=0