Update patch set 1
Patch Set 1:

We have to decide if we want to add both the announcement in one up-script or wait for the announce_before_call_setup one.

Additionally we have to decide if is fine to add the sounds under the "early_rejects" part and the preferences in the "Applications" section.

Patch-set: 1
Label: Code-Review=0