Update patch set 3
Patch Set 3:

Hello Rene,

I reviewed original issue 
https://support.sipwise.com/view.php?id=7227

Considering 
"In the CE documentation is written: "The {source, destination, direction} combination needs to be unique for a billing profile. The system will return an error if such a set is specified twice, both for the file upload and the input via the web interface." It does not respond with described above."

I'm not sure now if it is correct that we keep initial billing fee (lower id) in case when we met duplicates. Maybe we need to introduce some flag to allow user to control if he wants to overwrite existing, or keep old records.

Patch-set: 3