Update patch set 1
Patch Set 1:

why preparing backports first place before the initial commit passes the review ?
normally once the initial commit (branch: master) passes the review and it is merged - backports are created, especially if there are that many of them.

please change it to "modify" as that way mariadb threats it as an "in-column" change where "change" implies that you want to rename the column and change it's attributes. for this case there is no difference but you will have issues when using "change" over "modify" on auto_increment columns. hence I would like to keep the db.up scripts with the proper syntax first place, as other developers often just create db.up scripts from already existing ones and I do not want to repeat the same again.

Patch-set: 1
Label: Verified=0