Update patch set 3
Patch Set 3:

> I am regret this will brake replication dramatically if locations
 > are still in mysql. Isn't it?
 > P.S. I have NOT good option here...

I agree, but to TO BE is to have all the customers with location under REDIS. I added a note in the documentation in the upgrade pre-check steps.
This change in mysql db is just to keep everything aligned.

Patch-set: 3
Reviewer: Gerrit User 1000021 <1000021@85315e57-5f03-49c3-83f8-201ae2313a75>