TT#70327 Add 'stop_forking_code_lists' subscriber's preference
It contains a list of SIP codes that is matched with the response
code received in reply from the subscriber's contacts in case
of call failure. In case of match, the forking based on Q value
is stopped and the code is transparently signaled back to the
caller, whereas also an announcement can be played if configured
in the existing sound set. Response codes 600, 603, 604, 606 are
implicitly included in the list.

Change-Id: Iec6ccad25aa59130ecd5c9add2e83c1341ff49eb