Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> why csv should be backward compatible ?!
> customers can adapt their existing csv files they use for upload download with ease.

we always took care to keep it like this: with new ngcp version, a customer is still able to upload his existing fee .csv. there is a chance the customer does not even notice an error, but silently corrupting the fee records in the db. even worse, re-uploading adjusted .csv does not help then as it became part of fees_history. this can be avoided easily for this change.

anyway, if this is merged right now, rating (lcr, prepaid, rate-o-mat) will be broken unless they are adopted as well. as usual, we need to merge all at once then.

Patch-set: 4
Reviewer: Gerrit User 1000013 <1000013@85315e57-5f03-49c3-83f8-201ae2313a75>