Update patch set 4
Patch Set 4:

the proper approach is to add it into the release notes that the format of the column is changed.
so ones using it in the API can adapt to it.
The old format must not be supported and return an error (good point by the way!). so that we avoid silent errors and customers are forced to update their APIs in that case.
I know that is not best for existing customers and not something I am very fond of, but it that is to keep the database in the good state, also backward compatible.
You are right, this cannot be merged until everything is adapted first and then merged all together when everybody is ok with the changes.

As for taking care of the backward compatible changes, yes, we must have the db backward compatible when possible, means data can be converted back when needed. But having changes like that, is called "versioning", otherwise we will choke ourselves with limitations making things more complicated and not flexible. 
That is by the way how the world evolves :)

Patch-set: 4
Reviewer: Gerrit User 1000004 <1000004@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=0