Update patch set 4
Patch Set 4:

for api i did not respond since i agree. there is a proper error response in case (or easy to add), and it only affects one entry at a time. only minority of customer use the api to manipulate fees. wether acceptable to force customers to api adoptions in such case or not, is secondary for me since i dont expect complaints in that regards.

the majority use the fee upload, that's why it needs the main attention.
.csv upload processes lots of records at once. it silently skips malformed items, if detected at all. current panel impl for this new part seem to not bail out. malformed data can persist permanently in the history tables. the .csv format is specified by configyml, so we also maintain our column defaults there accordingly. wether a customer is aware of new columns in the .csv's or not; they will ask us to fix the mistake - and for fees this is known for hassles. adding or renaming a column is ok, changing an existing data type is not. 
i just see no reason now to break conventions we have in place to exaclty avoid that.

using an enum for flexibility, wether needed or not, does not help if this flexibility does not extend to all the remaining existing columns. and finally, there cant ever be a "mid-peak" enum value.

Patch-set: 4
Reviewer: Gerrit User 1000013 <1000013@85315e57-5f03-49c3-83f8-201ae2313a75>