Update patch set 4
Patch Set 4:

As I wrote above, I am fine with your approach. There are obviously pros/cons for both of them, but I see a good benefit of using the boolean multi selection, for both interface level usages and when used in the code.

------------------------------------
So for instance with "use_free_time enum it is flexible to have one column but it wont efficiently cover cases when 

onpeak_use_free_time
offpeak_use_free_time

then in the future:

something_else_use_free_time
something_extra_else_use_free_time

and if you want to use free time for "onpeak" and "something_extra" but not for "offpeak" and not for "something_else, you can easily accomplish that by checking the required fields.
That would not be easily possible with the enum in that case.
------------------------------------

P.S. As for preserving our conventions, to some extend yes, but it must not be something that we choke and hold ourselves from better code/schema for the sake of having old designs and old conveniences in place. There are a lot of "historical" places in the code as well as in the db where we either: keep struggling, or improve for the better good, it's us too, that should enjoy our code when possible.

Patch-set: 4
Reviewer: Gerrit User 1000004 <1000004@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=0