Update patch set 4
Patch Set 4:

> (2 comments)
 > 
 > Yes this definitely requires updating everything that uses the
 > `use_free_time` column. One way to avoid this would be to leave
 > `use_free_time` intact (boolean) and add another column
 > `use_free_time_for` (or something like that), defaulting to `both`
 > or `all`. Less clean in the DB but it wouldn't break anything. Just
 > a suggestion. Up to you guys.

We already have a lot of "leftovers" in the db that are hard to understand/inconvenient to use, so from my side, I would much prefer that we do that properly in the db (that is our sandbox afterall), if that requires one time changes in the code it's ok, as altogether those changes are not complex and easy to review / "find/grep".
`use_free_time` and `use_free_time_for` will surely be hard to understand for anyone who is not involved into the change.

Patch-set: 4
Reviewer: Gerrit User 1000004 <1000004@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=0