Update patch set 5
Patch Set 5:

(1 comment)

There is an issue in the up script, please check the comment.
Since it is just in master I think you can edit the up script without bit issue.

In the existing VM, instaed, the value has to be update manually in provisioning.voip_preferences and also in kamailio.dom_preferences

Patch-set: 5
CC: Gerrit User 1000021 <1000021@85315e57-5f03-49c3-83f8-201ae2313a75>