Update patch set 2
Patch Set 2:

> Doesn't a hardcoded salted password defeat the purpose of having a
 > salt? :D

I do not see a big problem with that in this particular case, as if we want to move it to bcrypt this solution will do. Either way, I do not think that in production environments this script will even do anything, as normally customers change the default adminstrator password.

Patch-set: 2
Reviewer: Gerrit User 1000004 <1000004@85315e57-5f03-49c3-83f8-201ae2313a75>