Update patch set 1
Patch Set 1:

> we do not hotfix db changes unless it's a critical hotfix

I agree that it is not critical, but it is anyway necessary for the future upgrade to mr7.5. I have seen next release is planned for beginning of May. If not customer will upgrade before, then it is fine to keep things as is.

Patch-set: 1
Reviewer: Gerrit User 1000021 <1000021@85315e57-5f03-49c3-83f8-201ae2313a75>