Update patch set 1
Patch Set 1: Code-Review-2

Same as with the commit to mr7.5, -2 just to clarify if I understand correctly.

This renaming work because 15553_not_replicated was not released yet in mr7.5.3 or anywhere else, because it's to recent, but if not it would have already been applied in some releases and so it could cause problems?

So just to reassure that this is safe to do for customers that might be using mr7.5.3 :)

Patch-set: 1
Reviewer: Gerrit User 1000042 <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=-2