Update patch set 1
Patch Set 1: Code-Review+1

It was 15554 in trunk from the very beginning:

17:28:36 ✔ taurus:(master)~/sipwise/git/db-schema$  git log ./db_scripts/diff/15554_not_replicated.up
commit 6618e9be0cf65f69c53e7ba3fa93e941f169aa04
Author: Marco Capetta <mcapetta@sipwise.com>
Date:   Wed Jun 26 15:07:23 2019 +0200

    TT#23358 DST change in kamailio
    
    Change kamailio.location expires and last_modified columns
    from datetime to bigint
    
    Change-Id: I2baa7134fe84960f07efda1e621cf0508345e1b1
17:28:48 ✔ taurus:(master)~/sipwise/git/db-schema$ 

How it appears 15553 in mr7.5?

P.S. this review is exactly the reason why we blocked initial backporting/hotfixing it for mr7.5.3 :-)
Now we can fix it internally.
AFAIK, we have committed this to mr7.5 only -> so now fixing in mr7.5.4 as already branched.


LGTM, I believe it is a correct fix for not-yet released mr7.5.4/mr7.5 but my question above probably should be answered in a commit message.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+1