Update patch set 1
Patch Set 1:

> Patch Set 1:
...
> now because the review number 15553 is already used, 15553_not_replicated.up in mr7.5 (and now in mr7.5.4) started to break 15553.up causing the latter to be skipped.

Tnx it is clear.
IMHO the proper fix here is TO create a dummy commit in master to reserve ID there. IMHO :

17:39:16 ✔ taurus:(master)~/sipwise/git/db-schema$ ./next_id 
15604


And then move in mr7.5/mr7.5.4 to those ID you reserved. Otherwise the same trouble can be created later on.

P.S. it is always bad to move steps to the past as we are changing execution order.
We should move steps to the future if necessary. Nope?

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@85315e57-5f03-49c3-83f8-201ae2313a75>