Update patch set 1
Patch Set 1:

I was thinking exactly the same in the beginning but there are also downsides to that.
if we have a dummy one, we create a scenario where we have revisions with the same number but different content, rendering it impossible to apply in the future even under a different id because we wont know where it has been applied and where it has not been applied (with the only possibility to create a completely new one with selects that distinguish where it has been applied before or not). with using an old, still unused in master id is relatively "safe" from the perspective that when we claim next id we use the script and never reuse old one and that opens a possibility.

of course if we guarantee 20000% that this revision is surely for mr7.5 only and under no circumstances we need it in mr8.0+, we can go the globally reserved dummy id in mr8.0+ (/cc Marco to confirm that)

Patch-set: 1
Reviewer: Gerrit User 1000004 <1000004@85315e57-5f03-49c3-83f8-201ae2313a75>