Update patch set 1
Patch Set 1: -Code-Review

(using same comment in both reviews)

I think that Alex suggestion of reserving a new number in all releases (trunk and then use this number in the older releases) makes sense and is safe for the changes proposed, and I cannot imagine very well the scenario in Kirill's concerns about having this change in other releases -- for me, as long as the number in trunk is not already in use in other releases, it can be included in other releases without interferences?  But probably I'm missing something and you know better this kind of scenarios.

In any case, I will remove my negative vote, now that the issue is considered and there's no danger of submitting by mistake.

Patch-set: 1
Reviewer: Gerrit User 1000042 <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=0