TT#133301 Let 'voip_pbx_autoattendants.choice' be compatible with string literals
As the option to add default actions has been introduced for Auto-Attendant,
where we are using string literals as lables/values, we need to let the SQL
structure be compatible with this update.

Hence our customers will be able to set and then see the default action
in the Auto-Attendant menu for a subscriber.

Change-Id: I949d7b72c9211588389a64df9af49b0201587d87