TT#58659 fix transcode_dtmf/cn pref type
All "transcode" preferences are listed as type "0" in the DB except for
DTMF and CN which are listed as "1". The LUA/Kamailio scripts fail to
include tyoe "1" preferences in the list of flags given to rtpengine.
Change the type to "0" to fix.

Change-Id: I38f0d9f1394e07b190f1bd196ac85f456017ae6f
(cherry picked from commit 531cc5c14d9b42bf783e7c2c745430b50f9e2d8b)