TT#109104 new peer preference 'allow_lm_forward_loop'
By default calls that come from a peer group and are looped back
to the same peer group due to a subscriber location mappings 'forward'
entry are blocked to avoid loop creation.

This preference allows to override the default behavior and permit
the loop over the same peering group.
It has to be used with caution and activated only if necessary.

Change-Id: I5040204805c3b2a740250cec11a8e9914333f10b