MT#56542 Add support of custom mysql instance
For preparation of dumps (sql and json) it's useful to use separate
mysql instance which is run and handled manually.
So add support of such instance.
Add variable MYSQL_SOCKET which shows that we want to connect to custom
mysql instance via defined socket.
In such a case don't try to start a service mysql.

Change-Id: I3f40a3857125fc72158aabf277f892c7cd541bf6