MT#58277 New preference sip_ping_notify and sip_ping_notify_codecs_list
If preference 'sip_ping_notify' is set and the user fails to reply
to the NGCP SIP ping (OPTIONS), the system will
trigger a CSTA event if and only if the reply
code matches the 'sip_ping_notify_codecs_list'.

Reply codes 200, 403, 404, 405 and 486 are considered 'good'
replies and will never trigger a CSTA notification.

Change-Id: I9639fdd2abdfb6beeb935d3c4f32e3c32d43f947