MT#58683 Set flags 8 in dispatcher when peer is probed
When we add the peering server into dispatcher table,
we need to set his flags column to 8, in order to enable
the probing, and keep flags 0 in the other internal
dispatcher entries (lb,sems,asterisk).

Also we have to update all current 'flags' values in
dispathcer for peering server (with setid=100) with the
new value 8.

Change-Id: Ie8c07639b49dcf8963459bfe1c3cabc12233c6a4