MT#57609 Add tables required by the FS mod_callcenter
This commit introduces a list of tables required
by the mod_callcenter module of the FS:
- `kamailio.queues` - custom table which is meant
  for usage by our internal http(s) gateway
- `kamailio.agents` - table taken from the upstream
- `kamailio.members` - table taken from the upstream
- `kamailio.tiers` - table taken from the upstream

Database `kamailio` will be used to store tables,
since traditionally other components such as voicemail
server or SEMS are using it as well for storing own tables.
Hence, there is no need to create new database for FS.

P.S.: Master upstream sources of the `mod_callcenter`
were used to get the SQL schemas.

Change-Id: Id88f5b6f8d6dd3f5681034e080a763c9beaba1cf