MT#60515 rename hm inbound/outbound directions to a_inbound/a_outbound
inbound/outbound no longer represent the "full" direction and belong
only to A leg, so they are renamed to a_inbound/a_outbound accordingly.

The up-script is not replicated because while the first half of the
nodes is upgraded the other half shoud continue to work with the old
parameters.

Change-Id: I1ac49b44c439ae1ba00c0862968c58831c605279