MT#60515 fix 15813_not_replicated script
The change of the ENUM 'direction' column removing old values
and adding the new ones, creates problems to existing HM
records.

The records with direction 'inbound' or 'outbound' loose
their direction right after the ENUM is updated.

To fix the issue we have before to add the new values,
then migrate the enties, and finally remove all the values
not needed.

Change-Id: Ib350b6aaada2cdabccb01947f131d3389e1a5110